requirements.txt
pytesseract
PyPDF2
Pillow
fitz
sentence-transformers
faiss-cpu
scikit-learn
openai
python-dotenv
